refactor(ocr): isolate converter implementations into engine/converters
